--- src/otx/core/data/transform/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""Module for OTX data transforms."""
