+++ src/otx/recipe/detection/atss_mobilenetv2_tv.yaml
@@ -40,7 +40,7 @@ overrides:
       train_subset:
         batch_size: 8
         transforms:
-          - class_path: otx.core.data.transform.detection.MinIoURandomCrop
+          - class_path: otx.core.data.transform_libs.torchvision.MinIoURandomCrop
             init_args:
               sampler_options:
                 - 0
