+++ src/otx/recipe/detection/ssd_mobilenetv2_tv.yaml
@@ -46,7 +46,7 @@ overrides:
             #   - 0.5
             #   - 1.5
             # hue_delta: 18
-          - class_path: otx.core.data.transform.detection.MinIoURandomCrop
+          - class_path: otx.core.data.transform_libs.torchvision.MinIoURandomCrop
             init_args:
               sampler_options:
                 - 0
